requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
mysql-connector-python>=8.4.0
pydantic>=2.9.0
python-dotenv>=1.0.1
